requirements.txt
# pip install -r requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
boto3==1.16.63
botocore==1.19.63
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
configparser==5.0.2
confuse==1.4.0
crayons==0.4.0
eel
et-xmlfile==1.1.0
flake8==3.9.2
google-auth==1.34.0
google-auth-oauthlib==0.4.5
greenlet==1.1.0
gspread==3.7.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
jmespath==0.10.0
mccabe==0.6.1
numpy==1.21.1
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.7
packaging==21.0
pandas==1.3.0
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodome==3.10.1
pyflakes==2.3.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-amazon-sp-api==0.6.2
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.7
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
toml==0.10.2
ulid-py==1.1.0
urllib3==1.26.6
webdriver-manager==3.4.2
xlrd==2.0.1

python-dotenv